вынес url логины и пароли в отдельный класс DataTest
