+++ .idea/workspace.xml
@@ -4,14 +4,13 @@
     <option name="autoReloadType" value="SELECTIVE" />
   </component>
   <component name="ChangeListManager">
-    <list default="true" id="62e66551-7af4-4e2a-a0cb-523ac424f2e0" name="Changes" comment="update">
-      <change afterPath="$PROJECT_DIR$/src/test/java/apiTest/model/settings/office/SettingOfficesBodyModel.java" afterDir="false" />
-      <change afterPath="$PROJECT_DIR$/src/test/java/apiTest/model/settings/office/WorkTimeOffices.java" afterDir="false" />
+    <list default="true" id="62e66551-7af4-4e2a-a0cb-523ac424f2e0" name="Changes" comment="обновлены api тесты для настроек офиса, тело запроса на создание нового офиса вынесено в модель">
       <change beforePath="$PROJECT_DIR$/.idea/workspace.xml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/workspace.xml" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/build.gradle.kts" beforeDir="false" afterPath="$PROJECT_DIR$/build.gradle.kts" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/test/java/apiTest/testCase/login/LoginApiTest.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/test/java/apiTest/testCase/login/LoginApiTest.java" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/test/java/apiTest/testCase/settings/SettingOfficesApiTest.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/test/java/apiTest/testCase/settings/SettingOfficesApiTest.java" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/src/test/java/dataTest/DataTest.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/test/java/dataTest/DataTest.java" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/test/java/uiTest/pages/autorization/AuthPage.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/test/java/uiTest/pages/autorization/AuthPage.java" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/test/java/uiTest/pages/settings/SettingOfficePage.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/test/java/uiTest/pages/settings/SettingOfficePage.java" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/test/java/uiTest/pages/settings/SettingOrganizationStructurePage.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/test/java/uiTest/pages/settings/SettingOrganizationStructurePage.java" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/test/java/uiTest/testCase/login/FrontLoginTest.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/test/java/uiTest/testCase/login/FrontLoginTest.java" afterDir="false" />
     </list>
     <option name="SHOW_DIALOG" value="false" />
     <option name="HIGHLIGHT_CONFLICTS" value="true" />
@@ -125,8 +124,8 @@
     "vue.rearranger.settings.migration": "true"
   }
 }]]></component>
-  <component name="RunManager" selected="Gradle.SettingOfficesApiTest.createOfficeTest">
-    <configuration name="LoginActionsTest.authenticateTest" type="GradleRunConfiguration" factoryName="Gradle" temporary="true">
+  <component name="RunManager" selected="Gradle.FrontLoginTest">
+    <configuration name="FrontLoginTest" type="GradleRunConfiguration" factoryName="Gradle" temporary="true">
       <ExternalSystemSettings>
         <option name="executionName" />
         <option name="externalProjectPath" value="$PROJECT_DIR$" />
@@ -139,7 +138,7 @@
           <list>
             <option value=":test" />
             <option value="--tests" />
-            <option value="&quot;apiTest.testCase.login.LoginActionsTest.authenticateTest&quot;" />
+            <option value="&quot;uiTest.testCase.login.FrontLoginTest&quot;" />
           </list>
         </option>
         <option name="vmOptions" />
@@ -248,11 +247,11 @@
     </configuration>
     <recent_temporary>
       <list>
+        <item itemvalue="Gradle.FrontLoginTest" />
         <item itemvalue="Gradle.SettingOfficesApiTest.createOfficeTest" />
         <item itemvalue="Gradle.SettingOfficesApiTest.deleteOfficeTest" />
         <item itemvalue="Gradle.LoginApiTest.retrieveTokenTest" />
         <item itemvalue="Gradle.LoginKeycloakTest.authenticateTest" />
-        <item itemvalue="Gradle.LoginActionsTest.authenticateTest" />
       </list>
     </recent_temporary>
   </component>
@@ -281,7 +280,7 @@
       <workItem from="1699509143695" duration="12000" />
       <workItem from="1699521739313" duration="339000" />
       <workItem from="1699972653555" duration="203000" />
-      <workItem from="1700028973978" duration="13949000" />
+      <workItem from="1700028973978" duration="15607000" />
     </task>
     <task id="LOCAL-00001" summary="добавил отдельный класс с данными">
       <option name="closed" value="true" />
@@ -475,7 +474,15 @@
       <option name="project" value="LOCAL" />
       <updated>1699972830902</updated>
     </task>
-    <option name="localTasksCounter" value="25" />
+    <task id="LOCAL-00025" summary="обновлены api тесты для настроек офиса, тело запроса на создание нового офиса вынесено в модель">
+      <option name="closed" value="true" />
+      <created>1700126636611</created>
+      <option name="number" value="00025" />
+      <option name="presentableId" value="LOCAL-00025" />
+      <option name="project" value="LOCAL" />
+      <updated>1700126636612</updated>
+    </task>
+    <option name="localTasksCounter" value="26" />
     <servers />
   </component>
   <component name="TypeScriptGeneratedFilesManager">
@@ -556,7 +563,8 @@
     <MESSAGE value="добавлены тесты api в настройки офисов" />
     <MESSAGE value="обновление тестов api в настройки офисов" />
     <MESSAGE value="update" />
-    <option name="LAST_COMMIT_MESSAGE" value="update" />
+    <MESSAGE value="обновлены api тесты для настроек офиса, тело запроса на создание нового офиса вынесено в модель" />
+    <option name="LAST_COMMIT_MESSAGE" value="обновлены api тесты для настроек офиса, тело запроса на создание нового офиса вынесено в модель" />
   </component>
   <component name="XDebuggerManager">
     <breakpoint-manager>
