+++ src/test/java/uiTest/pages/autorization/AuthPage.java
@@ -6,10 +6,6 @@ import static com.codeborne.selenide.Selenide.element;
 
 
 public class AuthPage {
-    private static final String LOGIN_MANAGER = "manager";
-    private static final String PASS_MANAGER = "manager";
-    private static final String LOGIN_OPERATOR = "operator";
-    private static final String PASS_OPERATOR = "operator";
 
     private final SelenideElement userNameInput = element("[name=username]");
     private final SelenideElement passwordInput = element("[name=password]");
@@ -25,22 +21,6 @@ public class AuthPage {
 
     public SelenideElement getLoginButton() {
         return loginButton;
-    }
-
-    public String getLoginManager() {
-        return LOGIN_MANAGER;
-    }
-
-    public String getPassManager() {
-        return PASS_MANAGER;
-    }
-
-    public String getLoginOperator() {
-        return LOGIN_OPERATOR;
-    }
-
-    public String getPassOperator() {
-        return PASS_OPERATOR;
     }
 
     // метод ввода логина и пароля
